@@ -1,13 +1,13 @@
 {
   "name": "angular2-jsonapi",
-  "version": "5.1.1",
+  "version": "6.0.1",
   "description": "A lightweight Angular 2 adapter for JSON API",
   "scripts": {
     "build": "rimraf dist src/compiled && tslint src/**/*.ts && ngc",
     "lint": "tslint src/**/*.ts",
     "pretest": "npm run build",
     "test": "karma start",
-    "prepublish": "ngc",
+    "prepublish": "npm run build",
     "ngc": "ngc",
     "servetest": "npm run pretest & karma start karma.dev.conf.js"
   },
@@ -38,10 +38,6 @@
     "qs": "^6.5.1"
   },
   "peerDependencies": {
-    "reflect-metadata": ">=0.1.3",
-    "rxjs": ">=5.2.0"
-  },
-  "devDependencies": {
     "@angular/animations": "^4.4.3",
     "@angular/common": "^4.4.3",
     "@angular/compiler": "^4.4.3",
@@ -50,14 +46,30 @@
     "@angular/platform-browser": "^4.4.3",
     "@angular/platform-browser-dynamic": "^4.4.3",
     "@angular/platform-server": "^4.4.3",
+    "reflect-metadata": ">=0.1.3",
+    "rxjs": ">=6.2.2"
+  },
+  "optionalDependencies": {
+    "fsevents": "*"
+  },
+  "devDependencies": {
+    "@angular/animations": "^6.1.3",
+    "@angular/cli": "^6.1.4",
+    "@angular/common": "^6.1.3",
+    "@angular/compiler": "^6.1.3",
+    "@angular/compiler-cli": "^6.1.3",
+    "@angular/core": "^6.1.3",
+    "@angular/platform-browser": "^6.1.3",
+    "@angular/platform-browser-dynamic": "^6.1.3",
+    "@angular/platform-server": "^6.1.3",
     "@types/date-fns": "^2.6.0",
     "@types/jasmine": "2.5.43",
     "@types/lodash-es": "^4.14.6",
     "@types/qs": "^6.5.0",
     "@types/reflect-metadata": "0.0.5",
     "@types/selenium-webdriver": "^3.0.7",
-    "codelyzer": "~3.2.1",
-    "core-js": "^2.5.1",
+    "codelyzer": "~4.2.1",
+    "core-js": "^2.5.7",
     "coveralls": "^2.11.14",
     "jasmine-core": "^2.6.4",
     "karma": "^1.7.1",
@@ -70,14 +82,14 @@
     "karma-webpack": "^2.0.3",
     "reflect-metadata": "^0.1.3",
     "rimraf": "^2.6.2",
-    "rxjs": "5.4.2",
+    "rxjs": "^6.2.2",
     "sourcemap-istanbul-instrumenter-loader": "^0.2.0",
     "ts-loader": "^2.3.7",
     "tslint": "^5.7.0",
     "tslint-config-airbnb": "^5.3.0",
-    "typescript": "2.5.3",
+    "typescript": "2.9.2",
     "webpack": "2.3.3",
-    "zone.js": "^0.8.16"
+    "zone.js": "^0.8.26"
   },
   "engines": {
     "node": ">=0.8.0"
