@@ -1,2 +1,2 @@
 // tslint:disable-next-line:variable-name
-export const AttributeMetadata = Symbol('AttributeMetadata');
+export const AttributeMetadata = Symbol('AttributeMetadata') as any;
