@@ -1,17 +1,15 @@
 import { AttributeMetadata } from '../constants/symbols';
 import { AttributeDecoratorOptions } from '../interfaces/attribute-decorator-options.interface';
-import { DateConverter } from '../converters/date/date.converter';
 import * as _ from 'lodash';
+import { serialize } from '@angular/compiler/src/i18n/serializers/xml_helper';
 
-export function Attribute(options: AttributeDecoratorOptions = {}): PropertyDecorator {
+export function NestedAttribute(options: AttributeDecoratorOptions = {}): PropertyDecorator {
   return function (target: any, propertyName: string) {
     const converter = function (dataType: any, value: any, forSerialisation = false): any {
       let attrConverter;
 
       if (options.converter) {
         attrConverter = options.converter;
-      } else if (dataType === Date) {
-        attrConverter = new DateConverter();
       } else {
         const datatype = new dataType();
 
@@ -31,13 +29,13 @@ export function Attribute(options: AttributeDecoratorOptions = {}): PropertyDeco
     };
 
     const saveAnnotations = function () {
-      const metadata = Reflect.getMetadata('Attribute', target) || {};
+      const metadata = Reflect.getMetadata('NestedAttribute', target) || {};
 
       metadata[propertyName] = {
         marked: true
       };
 
-      Reflect.defineMetadata('Attribute', metadata, target);
+      Reflect.defineMetadata('NestedAttribute', metadata, target);
 
       const mappingMetadata = Reflect.getMetadata('AttributeMapping', target) || {};
       const serializedPropertyName = options.serializedName !== undefined ? options.serializedName : propertyName;
@@ -45,24 +43,29 @@ export function Attribute(options: AttributeDecoratorOptions = {}): PropertyDeco
       Reflect.defineMetadata('AttributeMapping', mappingMetadata, target);
     };
 
-    const setMetadata = function (
-      instance: any,
-      oldValue: any,
-      newValue: any
-    ) {
-      const targetType = Reflect.getMetadata('design:type', target, propertyName);
+    const updateMetadata = (instance: any) => {
+      const newValue = instance['_' + propertyName];
 
       if (!instance[AttributeMetadata]) {
         instance[AttributeMetadata] = {};
       }
-      instance[AttributeMetadata][propertyName] = {
-        newValue,
-        oldValue,
-        nested:false,
-        serializedName: options.serializedName,
-        hasDirtyAttributes: !_.isEqual(oldValue,newValue),
-        serialisationValue: converter(targetType, newValue, true)
-      };
+      if (instance[AttributeMetadata][propertyName] && !instance.isModelInitialization()) {
+        instance[AttributeMetadata][propertyName].newValue = newValue;
+        instance[AttributeMetadata][propertyName].hasDirtyAttributes = !_.isEqual(
+          instance[AttributeMetadata][propertyName].oldValue,
+          newValue
+        );
+        instance[AttributeMetadata][propertyName].serialisationValue = newValue;
+      } else {
+        const oldValue = _.cloneDeep(newValue);
+        instance[AttributeMetadata][propertyName] = {
+          newValue,
+          oldValue,
+          converter,
+          nested: true,
+          hasDirtyAttributes: !_.isEqual(newValue, oldValue)
+        };
+      }
     };
 
     const getter = function () {
@@ -71,19 +74,10 @@ export function Attribute(options: AttributeDecoratorOptions = {}): PropertyDeco
 
     const setter = function (newVal: any) {
       const targetType = Reflect.getMetadata('design:type', target, propertyName);
-      const convertedValue = converter(targetType, newVal);
-      let oldValue = null;
-      if (this.isModelInitialization() && this.id) {
-        oldValue = converter(targetType, newVal);
-      } else {
-        if (this[AttributeMetadata] && this[AttributeMetadata][propertyName]) {
-          oldValue = this[AttributeMetadata][propertyName]['oldValue'];
-        }
-      }
-
-      this['_' + propertyName] = convertedValue;
-      setMetadata(this, oldValue, convertedValue);
+      this['_' + propertyName] = converter(targetType, newVal);
+      updateMetadata(this);
     };
+
 
     if (delete target[propertyName]) {
       saveAnnotations();
@@ -93,6 +87,7 @@ export function Attribute(options: AttributeDecoratorOptions = {}): PropertyDeco
         enumerable: true,
         configurable: true
       });
+
     }
   };
 }
